[esp32] Move OTA downgrade-protection tests into existing esp32 test module
The new tests/unit_tests/components/test_esp32.py collided with the
existing tests/component_tests/esp32/test_esp32.py basename, which pytest
rejects (no package __init__.py) when collecting the whole tests/ tree.
Move the tests into the existing module.
